Apparently we rely on this implicit case behavior. Let's revert it for now.

+++ source/tools/matchbin_metrics.h
@@ -100,7 +100,7 @@ namespace emp {
       return static_cast<double>(emp::hash_combine(
         qhasher(a),
         thasher(b)
-      )) / std::numeric_limits<double>::max();
+      )) / std::numeric_limits<size_t>::max();
     }
 
   };
